Removed Previous Pipeline Scripts and Example
